S15.e.5c: extract operator.ts header render (1318 → 1279 LOC)
Seventh sub-slice of the S15.e operator.ts decomposition train.
Lifts renderHeader + healthSummary out of startDashboard:

cli/src/commands/operator/render/header.ts (~98 LOC):
  - renderHeader(ctx) — closure deps: sandboxes, clusterData, meshHealth,
                       isRefreshing, spinFrames, spinIdx, clusterName,
                       viewMode, header, screen
  - healthSummary(sandboxes) — pure helper

A thin wrapper in operator.ts preserves the call site in render().

Behavioral change: NONE. Bodies byte-identical to originals.

Cumulative S15.e delta:
  pre-S15.e:  2894
  e.1-e.5b:   1318  (-1576)
  e.5c:       1279     (-39)
  Cumulative: -1615. Dialogs (~480 LOC) remaining for cap.

Verification:
- tsc --noEmit clean / lint 20 warnings (unchanged) / build clean
- npm test: 454 pass / 2 skipped pre-existing

Audit: docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e5c.md
