+++ cli/src/commands/operator.ts
@@ -43,6 +43,7 @@ import {
   renderAGT as _renderAGT,
   renderAGTFull as _renderAGTFull,
 } from "./operator/render/security.js";
+import { renderHeader as _renderHeader } from "./operator/render/header.js";
 
 // ── Command ─────────────────────────────────────────────────────────
 
@@ -297,68 +298,23 @@ async function startDashboard(refreshInterval: number, kubeContext?: string, dev
 
   // ── Rendering ─────────────────────────────────────────────────────
 
-  function healthSummary(): string {
-    const total = sandboxes.length;
-    if (total === 0) return "{gray-fg}no agents{/}";
-    const h = sandboxes.filter((s) => s.health === "healthy").length;
-    const d = sandboxes.filter((s) => s.health === "degraded").length;
-    const x = sandboxes.filter((s) => s.health === "down").length;
-    const parts: string[] = [];
-    if (h > 0) parts.push(`{green-fg}${h} healthy{/}`);
-    if (d > 0) parts.push(`{yellow-fg}${d} degraded{/}`);
-    if (x > 0) parts.push(`{red-fg}${x} down{/}`);
-    return `${total} agent${total === 1 ? "" : "s"} (${parts.join(", ")})`;
+  // Thin wrappers — bodies extracted to operator/render/header.ts (S15.e.5c)
+  function renderHeader(): void {
+    _renderHeader({
+      sandboxes,
+      clusterData,
+      meshHealth,
+      isRefreshing,
+      spinFrames,
+      spinIdx,
+      clusterName,
+      viewMode,
+      totalEgressCount: totalEgressCount(),
+      header,
+      screen,
+    });
   }
 
-  function renderHeader() {
-    const now = new Date().toLocaleTimeString();
-    const spin = isRefreshing ? `{cyan-fg}${spinFrames[spinIdx]}{/} ` : "";
-    const ctx = `{gray-fg}${clusterName}{/}`;
-
-    // Cluster health indicator
-    let clusterTag = "";
-    if (clusterData) {
-      const readyNodes = clusterData.nodes.filter((n) => n.status === "Ready").length;
-      const totalNodes = clusterData.nodes.length;
-      const nColor = readyNodes === totalNodes ? "green" : readyNodes > 0 ? "yellow" : "red";
-      const apiTag = clusterData.apiReachable ? "{green-fg}●{/}" : "{red-fg}●{/}";
-      clusterTag = `${apiTag} API  {${nColor}-fg}${readyNodes}/${totalNodes}{/} nodes  │  `;
-    }
-
-    // Mesh health indicator
-    let meshTag = "";
-    if (meshHealth) {
-      const relayColor = meshHealth.relayReady ? "green" : "red";
-      const regColor = meshHealth.registryReady ? (meshHealth.registryReadyPods < meshHealth.registryPods ? "yellow" : "green") : "red";
-      const regCount = meshHealth.registryPods > 0 ? ` (${meshHealth.registryReadyPods}/${meshHealth.registryPods})` : "";
-      meshTag = `{${relayColor}-fg}●{/} relay  {${regColor}-fg}●{/} registry${regCount}  │  `;
-    }
-
-    const viewLabel = viewMode === "cluster" ? "{blue-fg}{bold}[CLUSTER]{/bold}{/}  │  " : "";
-    const title = ` ${spin}{bold}AzureClaw Operator{/bold}  │  ${ctx}  │  ${viewLabel}`;
-    const stats = `${clusterTag}${meshTag}${healthSummary()}  │  ${totalEgressCount()} domain(s)  │  {gray-fg}${now}{/}`;
-    const shortStats = `${healthSummary()}  │  {gray-fg}${now}{/}`;
-
-    // Measure visible width (strip blessed tags, account for double-wide emoji)
-    const visWidth = (s: string) => {
-      const plain = s.replace(/\{[^}]*\}/g, "");
-      // Each emoji/surrogate pair occupies ~2 columns
-      let w = 0;
-      for (const ch of plain) {
-        w += ch.codePointAt(0)! > 0xffff ? 2 : 1;
-      }
-      return w;
-    };
-    const maxW = (screen.width as number) - 2;
-    const full = title + stats;
-    if (visWidth(full) > maxW) {
-      header.setContent(title + shortStats);
-    } else {
-      header.setContent(full);
-    }
-  }
-
-  // Thin wrappers — bodies extracted to operator/render/security.ts (S15.e.5b)
   function renderSecurity(): void {
     _renderSecurity({ agentTable, sandboxes, securityStates, securityBox, agtPanel });
   }
